区块链/实验/Ex4/requirements.txt
python-bitcoinlib
requests
